cap-aggregator/requirements.txt
aniso8601==9.0.1
azure-storage-blob==12.17.0
channels-redis==4.1.0
daphne==4.0.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.2.0
django-extensions==3.2.3
django-redis==5.3.0
django-storages==1.13.2
greenlet==2.0.2
iso639-lang==2.1.0
pip-chill==1.0.3
promise==2.3
psycopg2-binary==2.9.6
pyopenssl==23.2.0
pytz==2023.3
python-dotenv==1.0.0
service-identity==23.1.0
shapely==2.0.1
text-unidecode==1.3
whitenoise==6.5.0
zope.event==5.0
